+++ src/Plugin/Field/FieldType/PretixDate.php
@@ -24,6 +24,7 @@ use Nicoeg\Dawa\Dawa;
  * @property string uuid
  * @property string location
  * @property string address
+ * @property DateTimeComputed registration_deadline
  * @property DateTimeComputed time_from
  * @property DateTimeComputed time_to
  * @property int spots
@@ -44,6 +45,16 @@ class PretixDate extends FieldItemBase {
     $properties['address'] = DataDefinition::create('string')
       ->setLabel(new TranslatableMarkup('Address'))
       ->setRequired(TRUE);
+
+    $properties['registration_deadline_value'] = DataDefinition::create('datetime_iso8601')
+      ->setLabel(t('Registration deadline value'))
+      ->setRequired(TRUE);
+    $properties['registration_deadline'] = DataDefinition::create('any')
+      ->setLabel(t('Computed registration deadline'))
+      ->setDescription(t('The computed registration deadline DateTime object.'))
+      ->setComputed(TRUE)
+      ->setClass(DateTimeComputed::class)
+      ->setSetting('date source', 'registration_deadline_value');
 
     $properties['time_from_value'] = DataDefinition::create('datetime_iso8601')
       ->setLabel(t('Time from value'))
@@ -90,6 +101,11 @@ class PretixDate extends FieldItemBase {
         'address' => [
           'type' => 'varchar',
           'length' => 255,
+        ],
+        'registration_deadline_value' => [
+          'description' => 'The registration deadline value.',
+          'type' => 'varchar',
+          'length' => 20,
         ],
         'time_from_value' => [
           'description' => 'The time from value.',